Move SoundConfiguration into SoundLibrary
