+++ src/main/java/org/orecruncher/dsurround/config/libraries/ISoundLibrary.java
@@ -2,9 +2,11 @@ package org.orecruncher.dsurround.config.libraries;
 
 import net.minecraft.sound.SoundEvent;
 import net.minecraft.util.Identifier;
+import org.orecruncher.dsurround.config.IndividualSoundConfigEntry;
 import org.orecruncher.dsurround.sound.SoundMetadata;
 
 import java.util.Collection;
+import java.util.Optional;
 
 public interface ISoundLibrary extends ILibrary {
 
@@ -12,4 +14,11 @@ public interface ISoundLibrary extends ILibrary {
     SoundEvent getSound(final Identifier sound);
     Collection<SoundEvent> getRegisteredSoundEvents();
     SoundMetadata getSoundMetadata(final Identifier sound);
+
+    boolean isBlocked(final Identifier id);
+    boolean isCulled(final Identifier id);
+    float getVolumeScale(final Identifier id);
+    Optional<SoundEvent> getRandomStartupSound();
+    Collection<IndividualSoundConfigEntry> getIndividualSoundConfigs();
+    void saveIndividualSoundConfigs(Collection<IndividualSoundConfigEntry> configs);
 }
