+++ src/main/java/org/orecruncher/dsurround/sound/SoundVolumeEvaluator.java
@@ -6,9 +6,10 @@ import net.fabricmc.api.Environment;
 import net.minecraft.client.sound.SoundInstance;
 import net.minecraft.sound.SoundCategory;
 import net.minecraft.util.math.MathHelper;
-import org.orecruncher.dsurround.Client;
+import org.orecruncher.dsurround.config.libraries.ISoundLibrary;
 import org.orecruncher.dsurround.gui.sound.ConfigSoundInstance;
 import org.orecruncher.dsurround.lib.GameUtils;
+import org.orecruncher.dsurround.lib.di.ContainerManager;
 
 /**
  * Special hook into the Minecraft SoundSystem.  This logic scales the volume of a sound
@@ -17,6 +18,8 @@ import org.orecruncher.dsurround.lib.GameUtils;
  */
 @Environment(EnvType.CLIENT)
 public final class SoundVolumeEvaluator {
+
+    private static final ISoundLibrary soundLibrary = ContainerManager.resolve(ISoundLibrary.class);
 
     private static float getCategoryVolumeScale(final SoundInstance sound) {
         // Master category already controlled by master gain so ignore
@@ -37,7 +40,7 @@ public final class SoundVolumeEvaluator {
         if (!(sound instanceof ConfigSoundInstance)) {
             volume *= getCategoryVolumeScale(sound);
 
-            var volumeScale = Client.SoundConfig.getVolumeScale(sound.getId());
+            var volumeScale = soundLibrary.getVolumeScale(sound.getId());
             if (volumeScale != 0f) {
                 volume *= volumeScale;
             }
